requirements.txt.txt
fastapi
uvicorn
openai
azure-storage-blob
azure-search-documents
pyodbc
python-dotenv
